split command-test-base out
